third_party/WebKit/LayoutTests/http/tests/media/media-source/mediasource-play-then-seek-back.html

<!DOCTYPE html>
<html>
    <head>
        <script src="/w3c/resources/testharness.js"></script>
        <script src="/w3c/resources/testharnessreport.js"></script>
        <script src="mediasource-util.js"></script>
        <link rel="stylesheet" href="/w3c/resources/testharness.css">
    </head>
    <body>
        <div id="log"></div>
        <script>

            mediasource_testafterdataloaded(function(test, mediaElement, mediaSource, segmentInfo, sourceBuffer, mediaData)
            {

                mediaElement.play();
                // Append all the segments
                test.expectEvent(sourceBuffer, 'updateend', 'sourceBuffer');
                test.expectEvent(mediaElement, 'playing', 'Playing triggered');
                sourceBuffer.appendBuffer(mediaData);

                function confirmPlayThenEnd()
                {
                    test.waitForCurrentTimeChange(mediaElement, function ()
                    {
                        assert_false(mediaElement.seeking, 'mediaElement is no longer seeking');
                        test.waitForCurrentTimeChange(mediaElement, function() { test.done(); });
                    });
                }

                function delayedPlayHandler()
                {
                    assert_greater_than(mediaElement.currentTime, 0.0, 'Playback has started.');
                    test.expectEvent(mediaElement, 'seeking', 'mediaElement');
                    test.expectEvent(mediaElement, 'seeked', 'mediaElement finished seek');
                    mediaElement.currentTime = 0.0;
                    assert_true(mediaElement.seeking, 'mediaElement is seeking');
                    assert_equals(mediaElement.currentTime, 0.0, 'Current time is 0.0');

                    test.waitForExpectedEvents(confirmPlayThenEnd);
                }

                test.waitForExpectedEvents(function()
                {
                    test.waitForCurrentTimeChange(mediaElement, delayedPlayHandler);
                });

            }, 'Test playing then seeking back.');
        </script>
    </body>
</html>
